Reorder type machinery into more modules
This is only moving things around (and altering some doc comments), no
functional changes.
